+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-2018-07-16 - lua-ssllabs 0.2-0 - Sebastian Hübner <sh@kokolor.es>
+2019-02-01 - lua-ssllabs 0.2-0 - Sebastian Hübner <sh@kokolor.es>
+  * moved conatants tables DEFAULTS and ERRORS out of functions
+  * shortened function bool_to_paramter
+  * rewrote the api_requests function
+  * added two new public functions `ssllabs.from_cache()` and `ssllabs.new_assessment()`
+  * some code cleanup and typo fixes
+  * updates the documentation
+
+2018-07-16 - lua-ssllabs 0.1-2 - Sebastian Hübner <sh@kokolor.es>
   * added check_options functions which check the parameters
   * fixed a bug with default values in analyse function